cast int64

+++ internal/csl/builtins.go
@@ -105,7 +105,7 @@ func init() {
 			return nil, csl.ErrBadType
 		}
 		if runtimeNum == 0 {
-			return -1, nil
+			return int64(-1), nil
 		}
 		return int64(sizeNum / (runtimeNum * 60)), nil
 	}, nil)
